requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
google-generativeai
python-dotenv
flask
gunicorn
